added new component for renews
