+++ src/applications/mhv/medications/sass/medications.scss
@@ -38,7 +38,7 @@
 .refill-loading-indicator {
   margin-top: 70px;
 }
-.refill-pagination-container {
+.renew-pagination-container {
   width: 100%;
   display: flex;
   justify-content: center;
